refactor: rename post-transfer-pair.ts to post-transfer.ts for consistency
