+++ src/server/routes/transfers/index.ts
@@ -1,3 +1,3 @@
 export * from "./get-transfers";
-export * from "./post-transfer-pair";
+export * from "./post-transfer";
 export * from "./delete-transfer";
